Did a little code reorg
